# `ByteHelpers.ToHex`: Use .NET implementation